ZIRC: extract diffLeaves into formHelpers (single source)
diffLeaves was copy-pasted in six places (the five page editors +
SchemaForm) in three whitespace-variant bodies — all functionally
identical. Moved to one exported function in formHelpers.ts; all six
now import it.

Behavior-preserving, with one hardening: the key-union step no longer
iterates a Set (the project's tsconfig targets es5, where Set iteration
needs downlevelIteration). Deduping via concat+filter is target-agnostic
and produces the same key set in the same order.

First step of the ZircEntityEditor consolidation — see
reference/zirc-entity-editor-design.md.

ESLint clean; 49/49 ZIRC JUnit tests pass.
