+++ home/javascript/react/zirc/api/formHelpers.ts
@@ -40,3 +40,50 @@ export function seedFromDto<T extends object>(dto: T): FormFor<T> {
     }
     return out as FormFor<T>;
 }
+
+/**
+ * Recursively diff two form-data trees, returning one
+ * [jsonPointerPath, newValue] entry per changed leaf. The autosave loop
+ * walks the result and emits one field-path PATCH per entry.
+ *
+ * <p>Arrays are treated as leaves: a changed array yields a single entry
+ * for the whole array (compared by JSON.stringify), never per-element
+ * paths. That's deliberate — the server's FIELDS dispatch takes a whole
+ * array value for list fields (publications, reasons), and
+ * server-managed object arrays are filtered out upstream via the
+ * `managesOwnPersistence` skip-set before they ever reach here.
+ *
+ * <p>Plain objects recurse, prefixing the path. Leaf scalars compare by
+ * {@code Object.is}. A null/undefined new value is normalized to null on
+ * the wire.
+ */
+export function diffLeaves(
+    prev: unknown,
+    curr: unknown,
+    basePath = '',
+): Array<[string, unknown]> {
+    const isPlainObject = (v: unknown): v is Record<string, unknown> =>
+        typeof v === 'object' && v !== null && !Array.isArray(v);
+    if (Array.isArray(prev) || Array.isArray(curr)) {
+        if (JSON.stringify(prev ?? null) !== JSON.stringify(curr ?? null)) {
+            return [[basePath || '/', curr ?? null]];
+        }
+        return [];
+    }
+    if (isPlainObject(prev) && isPlainObject(curr)) {
+        // Union of keys, deduped without iterating a Set (keeps the helper
+        // target-agnostic — the project's tsconfig targets es5).
+        const keys = Object.keys(prev)
+            .concat(Object.keys(curr))
+            .filter((k, i, all) => all.indexOf(k) === i);
+        const changes: Array<[string, unknown]> = [];
+        for (const key of keys) {
+            changes.push(...diffLeaves(prev[key], curr[key], `${basePath}/${key}`));
+        }
+        return changes;
+    }
+    if (!Object.is(prev, curr)) {
+        return [[basePath || '/', curr ?? null]];
+    }
+    return [];
+}
